add osm_switch

+++ project/backend.py
@@ -148,6 +148,7 @@ class SkyML(Backend):
         img_provider: str,
         img_api_key: str,
         poly_dict: dict,
+        get_osm: bool,
     ):
 
         # set this to be used in task creation
@@ -194,15 +195,21 @@ class SkyML(Backend):
                     f"docker run --rm -v {remote_dir}:/output 316880547378.dkr.ecr.us-east-1.amazonaws.com/xv2-inf-backend:latest python backend_runner.py imagery --provider {img_provider} --api_key {img_api_key} --job_id {job_id} --image_id {img_id} --coordinates '{json.dumps(poly_dict)}' --out_path /output --pre_post {pre_post}"
                 )
 
+            # create list for extra args
+            inf_xtra_args = []
+
             # get OSM polygons
-            self._make_task(
-                f"docker run --rm -v {self.remote_poly_dir}:/output 316880547378.dkr.ecr.us-east-1.amazonaws.com/xv2-inf-backend:latest python backend_runner.py fetch_polys --job_id {job_id} --coordinates '{json.dumps(poly_dict)}'"
-            )
+            if get_osm:
+                self._make_task(
+                    f"docker run --rm -v {self.remote_poly_dir}:/output 316880547378.dkr.ecr.us-east-1.amazonaws.com/xv2-inf-backend:latest python backend_runner.py fetch_polys --job_id {job_id} --coordinates '{json.dumps(poly_dict)}'"
+                )
+                
+                inf_xtra_args.append("--bldg_polys /input/polys/polys.geojson")
 
             # run xv2
             self._make_task(
                 # Todo: currently skips using bldg_polys
-                f"docker run --rm --gpus all --shm-size 56g -v {self.remote_pre_in_dir}:/input/pre -v {self.remote_post_in_dir}:/input/post -v {self.remote_temp_out}:/output -v {self.remote_poly_dir}:/input/polys 316880547378.dkr.ecr.us-east-1.amazonaws.com/xv2-inf-engine:latest --dp_mode",  # BUG: Bug in inference engine does not produce correct outputs with 4 GPUs unless run in dp_mode. Adding flag as stopgap
+                f"docker run --rm --gpus all --shm-size 56g -v {self.remote_pre_in_dir}:/input/pre -v {self.remote_post_in_dir}:/input/post -v {self.remote_temp_out}:/output -v {self.remote_poly_dir}:/input/polys 316880547378.dkr.ecr.us-east-1.amazonaws.com/xv2-inf-engine:latest --dp_mode {' '.join(inf_xtra_args)}",  # BUG: Bug in inference engine does not produce correct outputs with 4 GPUs unless run in dp_mode. Adding flag as stopgap
                 gpu=True,
             )
